Capped datasize in obj.data
The data size is now capped at 235bytes  or 1880 bites in obj.data. This
will allow our class to continue running even if it has an error in data
size that would require it to call for an index that does not exist.
I also added the description at the top of this class.