Illegitimate frame properties will result in error
If you create m = FrameObj(ACK, x, y, don't care) and call m.data or
m.CRC8 there will be an error as an ACK frame contains neither data, nor
does it need to check the correctness of the data.
dataSize is a awkward case, I left it as is, ack.dataSize =0 but it may
make more sense to un comment the error instead.
I also added some whitespace for legibility